requirements.txt
Flask
uuid
flask-restful
python-dotenv
openai-whisper
pydub
numpy